+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog], and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
+### Fixed
+- `core` - Fix helper data parsing in `getRecoveredData` method
 
 ## [0.3.0] - 2025-08-22
 ### Added