+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog], and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
+
+## [0.4.0] - 2025-09-15
 ### Added
 - `core` - Add `factors` and `walletAddress` SDK options.
 - `core` - Add location hash compose/parse utils.
@@ -40,7 +42,8 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
-[Unreleased]: https://github.com/rarimo/unforgettable-sdk/compare/0.3.1...HEAD
+[Unreleased]: https://github.com/rarimo/unforgettable-sdk/compare/0.4.0...HEAD
+[0.4.0]: https://github.com/rarimo/unforgettable-sdk/compare/0.3.1...0.4.0
 [0.3.1]: https://github.com/rarimo/unforgettable-sdk/compare/0.3.0...0.3.1
 [0.3.0]: https://github.com/rarimo/unforgettable-sdk/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/rarimo/unforgettable-sdk/compare/0.1.0...0.2.0